del: suite-test indexing
The indexing of suite tests was motivated to have a unique suite-test
identifier for concurrency save suite-test specific fixture creation in
a suite's SetUp-method.  Since the SetUp-Method is provided with the
gounit.T pointer of the suite-test for which it is called and this
pointer is unique for each suite test we can use this pointer as
identifier for a suite-test's fixture.

Removing the indexing decreased the execution time of the test-suites of
this project (so far) by more than 50%!