del: lines: concurrency safety
Even in this early stage having concurrency safety for a view turned out
to be rather complex since everything which can be retrieved/changed of
a view and its components needs to be protected by mutexes. On the other
hand if a view is only and only changed as consequence of an reported
event and the manipulation of the view stops when the reporting call back
returns then the serialization through the event-polling suffices.
Because view internal screen synchronization always happens after the
reporting call back returns.  For the use-case that a client wants to
change the views state without a user triggered input event, i.e. a file
on disk has changed, an Update-method will be provided which posts an
update event into the event-queue that in turn will be processed through
regular event-polling.