feat: register update listener
To avoid data races a view should be only manipulated during the
execution of a listener callback.  Such a listener callback is usually
triggered by a user input.  But there are also use case were a client
wants to change a view without user-interaction, i.e. a file on disk
changed.  For this use case  an update-event with its listener can be
created by registering an update listener.  This  posts an update event
into the event queue which once polled calls back to the update
listener.  In this way we stay save in the "event" - "report" -
"screen-sync" circle.