refac: lines: overall architecture
Initially lines was designed around a View-Type putting the View/Screen
in the center.  At the same time it was attempted to relive a user of
screen synchronization and event-loop management.  Since the screen
synchronization investigates the whole view and if there where updates
also changes its state a user shouldn't modify the View during this
process which is hidden from the user, to avoid race conditions.  Making
the view concurrency save would have even in this early stage added a
fair amount of complexity and overhead to the design.  A more appealing
design seemed to be to design *lines* around the event handling and
provide the event listeners with a view-instance to manipulate.  Thus if
the listener callback has finished all its manipulations before
returning and the user doesn't keep a view around it is guaranteed
through the serializing event loop polling that the screen
synchronization is not interfered by other operations.  Therefor if the user
is not modifying the view concurrently and screen synchronization isn't
as well race-conditions are avoided by design.