fix: race-conditions in test fixture
The event-loop of a wrapped Register instance for testing is now by
default run in its own go routine.
It is now guaranteed (by running all tests for 1000 times having the
race detector turned on) that an method of the Register-fixture that
triggers an post to the event-loop returns after that event was fully
processed.