feat: Keyboard event in favor of Runes event
Runes allowed the user only to take control over the runes input
Keyboard reports to its listener also the special keys and modifiers
shadowing all other keyboard event listeners except for the quit event.
I.e. the user can now take over the keyboard if wanted.  The main
problem of the implementation was how to end the exclusive keyboard
listening.  The seemingly most direct solution is to set in the listener
callback the Keyboard listener to nil.  Since the registration of events
is implemented concurrency save the retrieval of an event listener must
be separated from its execution to avoid dead locks.