refac: *Keys*-type became *Features*
This name resembles better what this type is for: To add/remove library
features associated with given feature triggering event keys.  For
example adding the features FtUp and FtDown enables up and down
scrolling.  While features have commonly used defaults for their
triggering event keys they may be changed in a consistent way.  If a
feature has the same event keys as a user defined key-event the users
listener is called first followed by the application of the associated
feature.