add: lines: storage for events-listeners mappings
The Listeners type moves the concurrency save event listener storage out
of the Register-type whose main job is to manage the event loop,
reporting of events and synchronization of the screen.  I.e. the
implementation of Register will be less cluttered.  It will also make it
easier in the future to create several sets of listeners registration which
may be bound to various View-components or "contexts".

In preparation to factor the concurrency save key/rune
event-registration storage out of the Register type.  This will make it
easier in the future