refac: lines: Listeners-integration
Factored the events-listeners mappings from the Register type out to the
Listeners-type.  With introducing the Features concept and adding the
idea of ViewComponet as interface of functions ID() uint64,
Parent() ViewComponet and Focused() ViewComponent we can map a
ViewComponent to a pair of Features and Listeners.  If then an event
occurs we can ask a View which also implements the ViewComponent
interface for its focused component can see if it is associated with a
Features-Listeners pair and report the event and bubble up over the
parents to the root element, i.e. View-instance.