add: lines: Events: Reported-event
Events provides now the option to register a listener for the "reported"
event.  The reported event reports back when an event was reported
(except for the reported-event).  This might be useful in the future for
logging as of now it simplified the implementation of the automatic
event-loop termination after a given number of reported event in the
Events-fixture tremendously.  Now there are only three Events-methods
left which need to be shadowed by the fixture-implementation:
Listen, QuitListening and Update.  Since all three of them are
synchronizing:
- initial resize event
- stop polling
- update event reported
non of them can be refactored away.