refac(lines): add testing-features to production
The Events-fixture type was implemented in the "lines_test" package and
thus not available for testing outside the lines package.  I.e.
packages who use this package can not profit from the implemented
testing features like firing an event or retrieving the screen's
content.  To make these features available a Testing-type and a Test
function has been added and the Events-type has been connected with this
testing type to be aware of "testing".  I.e. instead of defining an
interface and returning different implementations of Events to
facilitate testing the Test-constructor returns an Events instance and a
Testing instance whereas the former has a slightly different behaviour
than in production and the later provides the testing features.  The
changes are Listen is now non-blocking but is guaranteed to return after
the initial resize completed.  Update is guaranteed to return after its
generated Update-event and all events created inside the update-callback
have completed.  QuitListening returns after the quit process has
completed and Testings LastScreen-property was set.  Since starting and
ending listening are happening only once checking for testing doesn't
generated overhead.  And since the typical use case of an Update call is
inside an heavy operation the two checks for testing also wont generate
any noticeable overhead.