improve: equality assertion
The equality assertion recognizes now pointers and treats them according
to their "==" definition in go.  Stopped using cmp.Diff as default
decision tool for equality in favor of "==" and focusing on strings
which seems more reasonable in a testing context.  I.e. if provided
arguments are not pointers it is investigated if they are strings,
Stringer or others.  In the two former cases the provided strings are
compared by "==" and cmp.Diff is only used if this comparison fails to
integrate an diff into the error message.  In the later case again
strings are created this time by fmt.Stringf("%v", x) and compared.
The inverse is done for the not-equal assertion.