test: for file system operations and their mocking
Added a tool set of potentially failing file system operation to the
FS-type which is propagated to all Dir and TmpDir instances to use it
for their file system operations.  These functions default to the
corresponding functions of the standard library.  In addition an
FS-fixture implementation was added which allows to mock up each of
these file system operations with failing functions as needed to easily
test the error handling.