refac: module: generalized to Sources
Since a go module may be huge we might want only to watch a source
directory inside a module.  Hence the Module type became the Sources
type which discriminates now the module directory and the watched
directory which may or may not be the module directory and which may or
may not be a go package.  The later case is useful if a directory is
watched which has nested packages but is not a package itself.