add: gounit-cmd: view
A new view is initialized with an Initer implementation providing the
initial setup for the view and getting in return the functionality to
modify the view.  I.e. there is no view-instance with methods to
manipulate it but rather a set of callback functions which provides the
most freedom on the client side and on the side of the view for how to
utilize respectively implement the view.  The disadvantage is that an
extension of the view "api" will most likely be a braking change.