refac: pkg/module => cmd/gounit/model
It turned out that Module represents the gounit's command data
structure.  Also the use of the Module package outside gounit seems
rather unlikely.  Hence it makes more sense to have it not in gounit's
package directory but to make it the model package of the gounit
command.  We finally arrived naturally by the good old MVC-architecture.
Not the nonsense which can be found at Wikipedia but an MVC-architecture
were view and model are completely isolated from other components of the
program and only care about/responsible for the API they provide.  These
APIs are then used by the controller to display information from the
model through the view to the user as requested by the later.