refac: controller: locked view update
The view can receive updates from at least two "channels": user input
and programmatically reported test-results.  To avoid race conditions
the initial idea was to send updates for a view through a channel of a
concurrently running function executing the update.  This makes it hard
to test since we dont know when a simulated user-input had its effect on
the screen.  Replacing the channel with a mutex-protected Update
function has the same effect with the advantage that this function
doesn't return before the view manipulation is finished (an property of
lines event-handling during testing).