git: ignore go.mod/go.sum of goldenmod
In the gounit command's controller's testdata directory is a golden
module for executing end-to-end tests.  In order to execute this
module's package tests a go.mod and go.sum file is needed which is
automatically created by the controller fixture setup functionality.
Hence it doesn't need to be part of the version control.