refac: buttons; stacking reporting content
Buttons: have been simplified by removing the package view and the suite
variant.  If we in addition remove the args button there are three
buttons removed which can be replaced by vet, race and stats buttons.
The functionality represented by the "package view" and the "suite
variant" are no implicit provided through selecting a displayed package
name or suite name.

stacking reporting content: Even though I'm advocating to not abuse a
view as data storage it seemed to be more reasonable for now to add this
feature to the view instead of caching a "currently reported content" in
the controller in case the client requests the help or about screen.
I.e. a reporting content may be pushed and popped now in case it is
desired to go back to a specific view.  At the end this also doesn't
seem satisfying: It probably is best having "ReportedTests"-instance in
the controller which provides a reportable liner of its current state if
asked.  I.e. we go back to a unstacked view an if we go back from the
help or about display we ask the ReportedTests-instance for its current
state to pass it on to the view.