fix: init golden: initialization race
Once per test run the go.mod and go.sum file of the golden module in
controller/testdata should be (re)created.  The natural location for
such a doing would be the the package init() function.  But since we
have quite a few lines of code which are involved with that subject
(creating and caching of go.mod and go.sum) which depend on Testing
instance, that was not an option without rewriting this code.  Hence now
each suite using the golden module calls the function to initialize it,
which is an no-op iff it has been already called and which is
concurrency save.