feat: packages/suites only are selectable
Ticking with j/k or up/down keys through the reported lines allows only
lines to select which are either flagged a package or a suite line.
Other lines are not selectable with the key-board.  While a shift click
on a test will show its file location which doesn't make sense to
support for keyboard selection.