refac: controller: model-state reporting
There is now a simple *report* type to report things like the help or
about screen.  Then for reporting model-states there are more complex
types embedding *report* like goReport for reporting a testing package
from the model which has only go-tests.  To control the model state
reporting there is a reporter interface summarizing the commonalities of
the more specific report-types.  For transitioning between different
views of the reported model-state the more complex report-types are
typed by a reportType-property.  E.g. if a reported package-line was
clicked then the list of packages is shown if the current report is not
of type rprPackages otherwise the clicked package is reported.