refac: controller: model-state reporting
Discarded the idea of "more complex"-report types embedding a basic
report type in favor of an basic report type extended by a type property
and specialized functions which calculate lines of an report.  I.e. a
combination of these functions calculate a reports data set.  This
approach seems to promise more flexibility.