feat: controller: report latest/locked (go) suite
The default reporting has changed now: there is now the last suite of
the latest modified test-file shown.  In case there are go-tests or more
than one suite in a test-file and the use selects instead of the default
suite and other suite or go-tests (go-suite) this selection is kept over
file updates.  Unless of course there is a failing test in a not
selected (go) suite.