refac: controller: model-state: minimize locking; unbrittle reporting
The original idea was to clone the currently reported model-state of the
controller if it needs updating due to a source-change or the reported
aspect changed due to a user input.  To avoid race conditions between
user input and source-changes the locking ended up spanning the whole
report calculation process which is probably the most expensive process
in the controller.  Now the locking is reduced again to the cloning of
a controller's model-state and its reset while the report calculation is
done on the cloned state.  For the pathological case that a user-input
reveres to an updated or stale state reporting now falls back to a
packages/package-folded report if inconsistencies occur respectively the
report is discarded when it's based on a stale state.