feat: switch vet and race detector on
Is vet/race switched on while a particular package is shown the tests of
this package are rerun with the corresponding arguments set.  Has the
user selected the packages overview (i.e. a list of all packages folded
is reported) and switches vet/race on and updates the source code of a
watched package or selects a package to report the tests of this package
are (re)run with corresponding arguments set.