refac: made "more"-buttons the default buttons
and put the switches in a dedicated "switches" button bar.  It seemed
to make more sense having the "help" and "quit" button available if
you especially start gounit the first time.  Since it is not necessary
to go "back" once the "switches"-button was selected it seems also not
to much effort to select "switches" once per session.