fix: controller: empty report data race
The empty report was defined as constant.  Since the listener of a
report is updated by the controller's model state before it is reported
to the view we create a race condition in this way.  We could also
define the empty report model-state specific and pass a model state's
empty report on to the report constructor to return it in the "empty"
case.  Since the empty report is a rare exception it seems not worth to
add a further property to the model state and a further argument to the
report constructor.