feat: panics are reported per test
So far a packages test run stopped at test-panic while the whole package
test run was considered as failed reporting the panic's output.  While
possible it made it quite tedious to figure out the panicking test.  I
case the panic occurred in an other go routine then the one which is
executing the test it was impossible to figure the panicking test.  No a
test panic is reported as such whereas the panicking test is reported
at the beginning followed by the panic message followed by the code file
and line where the panic occurred followed by the panic's stack-trace.

Note while the reporting improved this doesn't solve necessarily issue
when a go routine panics which is not the one executing the test.  But
if that go routine were to propagate the panic to the testing go routine
like it is for example done in

    github.com/slukits/lines/internal/term/transactions.go

then it is possible to concatenate two panic traces reported at the
executing test.